requirements.txt
# Flask Core
Flask==2.3.3
Werkzeug==2.3.7

# Database
Flask-SQLAlchemy==3.0.5

# Forms
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0

# Email
Flask-Mail==0.9.1

# Image Processing
Pillow==10.0.0

# Environment Variables
python-dotenv==1.0.0

# Admin Interface (Optional)
Flask-Admin==1.6.1

# Security
Flask-Talisman==1.1.0

# Development Dependencies
Flask-DebugToolbar==0.13.1

# Production Server
gunicorn==21.2.0

# Additional Utilities
python-dateutil==2.8.2
bleach==6.0.0